requirements.txt
# Quantum computing
qiskit>=1.0,<2.0
qiskit-aer>=0.13
qiskit-machine-learning>=0.7

# Classical ML
scikit-learn>=1.3
scikit-optimize>=0.9

# Optimization
cvxpy>=1.4

# Data
numpy>=1.24
pandas>=2.0
scipy>=1.11

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Config
pyyaml>=6.0

# Notebooks
jupyter>=1.0
ipykernel>=6.0

# Testing
pytest>=7.0
